refactor: streamline feature organization and update references to match new structure
- Removed unused `connections` and `dragon-fight` feature loaders.
- Renamed and reorganized files for improved clarity (e.g., `health-manager` to `dragon-fight-health`).
- Updated file imports in `events/index.ts` and other modules to reflect new paths.
