--- behaviour_pack/scripts-dev/features/connections/index.ts
@@ -1,7 +1,0 @@
-import handleConnections from "./connections";
-import handleWhitelist from "./whitelist";
-
-export default function loadConnectionsFeature(guild_id: string) {
-    handleConnections()
-    handleWhitelist(guild_id)
-}
