--- behaviour_pack/scripts-dev/features/dragon-fight/index.ts
@@ -1,5 +1,0 @@
-import loadHealthManager from "./health-manager";
-
-export default function loadDragonFightFeature() {
-    loadHealthManager()
-}
